use Lorenzo's gtkmm
svn path=/trunk/yarp2/; revision=8625

+++ packaging/windows/src/build_gtkmm.sh
@@ -29,16 +29,23 @@ fi
 
 fname=gtkmm-$BUNDLE_GTKMM_VERSION
 
-if [ ! -e $fname.exe ]; then
+ext=exe
+if [ ! "k$BUNDLE_GTKMM_ZIP" = "k" ] ; then
+  ext=zip
+fi
+
+if [ "$ext" = "exe" ]; then
+
+  if [ ! -e $fname.exe ]; then
 	GTKMM_DIR=`echo $BUNDLE_GTKMM_VERSION | sed "s/\.[-0-9]*$//"`
 	wget -O $fname.exe http://ftp.gnome.org/pub/GNOME/binaries/win32/gtkmm/$GTKMM_DIR/gtkmm-win32-devel-$BUNDLE_GTKMM_VERSION.exe || (
 		echo "Cannot fetch GTKMM"
 		rm -f $fname.exe
 		exit 1
 	)
-fi
+  fi
 
-if [ ! -d $fname ]; then
+  if [ ! -d $fname ]; then
 	mkdir -p $fname
 	cd $fname
 	7z x ../$fname.exe || {
@@ -47,9 +54,9 @@ if [ ! -d $fname ]; then
 		rm -rf $fname
 		exit 1
 	}
-fi
+  fi
 
-if [ ! -e $BUILD_DIR/$fname/include ]; then
+  if [ ! -e $BUILD_DIR/$fname/include ]; then
 	# rejigger files for Lorenzo
 	cd $BUILD_DIR/$fname
 	mkdir -p include || exit 1
@@ -65,7 +72,37 @@ if [ ! -e $BUILD_DIR/$fname/include ]; then
 	cp $ALT/include/glibconfig.h ../lib/glib-2.0/include || exit 1
 	mkdir -p ../lib/gtk-2.0/include
 	cp $ALT/include/gdkconfig.h ../lib/gtk-2.0/include || exit 1
+  fi
 fi
+
+
+if [ "$ext" = "zip" ]; then
+
+  if [ ! -e $fname.zip ]; then
+	wget -O $fname.zip $BUNDLE_GTKMM_ZIP || {
+		echo "Cannot fetch GTKMM"
+		rm -f $fname.zip
+		exit 1
+	}
+  fi
+
+  if [ ! -d $fname ] ; then
+    if [ ! -e gtkmm ] ; then
+      unzip $fname.zip || {
+	  	echo "Cannot unpack GTKMM"
+		cd $BUILD_DIR
+		rm -rf gtkmm
+		exit 1
+	  }
+	fi
+	#mv $fname.zip store-$fname.zip
+	#mv $fname.exe store-$fname.exe
+	mv gtkmm $fname || exit 1
+	#mv store-$fname.zip $fname.zip
+	#mv store-$fname.exe $fname.exe
+  fi
+fi
+
 
 (
 	GTKMM_DIR=`cygpath --mixed "$BUILD_DIR/$fname"`
